+++ src/main/java/com/javaguru/shoppinglist/console/WrongMenuInputException.java
@@ -1,4 +1,6 @@
-package com.javaguru.shoppinglist.domains.products.productService.validation.exceptions;
+package com.javaguru.shoppinglist.domains;
+
+import com.javaguru.shoppinglist.domains.products.productService.validation.exceptions.ProductValidationException;
 
 public class WrongMenuInputException extends ProductValidationException {
