--- src/main/java/com/javaguru/shoppinglist/domains/basket/Baskets.java
@@ -1,5 +1,0 @@
-package com.javaguru.shoppinglist.domains.basket;
-
-public class Baskets {
-
-}
