+++ src/main/java/com/javaguru/shoppinglist/domains/products/productService/validation/exceptions/ProductDiscountIllegalException.java
