+++ src/main/java/com/javaguru/shoppinglist/domains/userItems/userItemService/validation/exceptions/UserException.java
