+++ src/main/java/com/javaguru/shoppinglist/domains/userItems/userItemService/validation/rules/UserItemProductRule.java
